pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "dataup"
version = "0.1.0"
description = "Official Python SDK for the DataUp API"
readme = "README.md"
license = "MIT"
requires-python = ">=3.10"
authors = [
    { name = "DataUp", email = "support@dataup.io" }
]
keywords = ["dataup", "api", "sdk", "computer-vision", "ml", "inference"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Typing :: Typed",
]
dependencies = [
    "httpx>=0.25.0",
    "pillow>=12.1.0",
    "pydantic>=2.0.0",
    "click>=8.0.0",
    "rich>=13.0.0",
    "dataup-models",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=4.0.0",
    "respx>=0.21.0",
    "mypy>=1.8.0",
    "ruff>=0.2.0",
]
ultralytics = [
    "ultralytics>=8.3.0",
]
roboflow = [
    "roboflow>=1.0.0",
]
all = [
    "ultralytics>=8.3.0",
    "roboflow>=1.0.0",
]

[project.scripts]
dataup = "dataup.cli.main:cli"

[project.urls]
Homepage = "https://dataup.io"
Documentation = "https://docs.dataup.io/sdk/python"
Repository = "https://github.com/dataup-io/dataup-sdk"

[tool.hatch.build.targets.wheel]
packages = ["src/dataup"]

[tool.mypy]
python_version = "3.10"
strict = true
warn_return_any = true
warn_unused_configs = true

[tool.ruff]
target-version = "py310"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B", "SIM"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"


[tool.uv.sources]
dataup-models = { git = "https://github.com/dataup-io/dataup-models.git", rev= "v1.1" }
